샐러리 수정중

+++ backend/src/main/java/erp/backend/domain/salary/entity/Salary.java
@@ -41,8 +41,8 @@ public class Salary {
     private double salaryPayMoney;
 
     @Column(name = "SALARY_TAX")
-    private int salaryTax;
+    private double salaryTax;
 
     @Column(name = "SALARY_BONUS")
-    private int salaryBonus;
+    private double salaryBonus;
 }
